switch to class-per-camera and not per robot
this makes it easier to have more than two cameras on the robot and just
in general really really easy to extend the vision system to more
cameras. additionally, it prevents mistakes as you only have to change
code in one place rather than in n places where n is the number of
cameras on the robot.